[Libcast-in-Chrome] Expose Session Capabilities
Currently, the SenderSession assumes that we only want to
check capabilities when we have already decided to start
remoting. However, after work to consume this API in Chrome
it became obvious that we actually want the capabilities first,
so we can decide if we even want to do remoting at all.

This patch exposes methods to request capabilities, get them
in a callback, as well as cleans up the Negotiate() structure
since we don't need duplicate OnNegotiated*() calls.

Bug: 1318499

5.2

Change-Id: Id39cf60f8b94c7b1a17f7b934adec99c3d6bde15