[Libcast-in-Chrome #5] Cleanup SenderSession API
Currently, the SenderSession assumes that we only want to
check capabilities when we have already decided to start
remoting. However, after work to consume this API in Chrome
it became obvious that we actually want the capabilities first,
so we can decide if we even want to do remoting at all.

This patch exposes methods to request capabilities, get them
in a callback, as well as cleans up the Negotiate() structure
since we don't need duplicate OnNegotiated*() calls.

This patch also includes some additional API cleanups, such
as adding more error types for Answer responses and adding the
ability to tell that a message has timed out.

Bug: 1318499
Change-Id: If22eaef6572607fa4757eba63848a98b9c2ee566