[MaaF] Rename PipelineStatus message to ErrorType
This name will more appropriately encase a wide variety of remoting
errors in the future. We can continue to add values to this enum in
sections (i.e, the next, non-media pipeline errors could start at 101, 102, ...)

Additionally, correct a typo and bring comment in line with the rest
of the file.

Change-Id: I3dca182c2f0e7b237cae093fc60a016f0e0ff7c0