[Open Screen] Replace ByteBuffer with a generic span type.
This replaces ByteBuffer with a generic span type that is generally
forward compatible with C++20 std::span.  span permits a view on both
const and non-const uint8_t data.

ByteView remains an alais for span<const uint8_t> and ByteBuffer is an
alias for span<uint8_t>.

Updates the public APIs for libcast to use the aliases and removes
remaining dependencies in public headers on absl::Span.

Removing absl::Span from the remainder of the codebase can be done as a
mechanical cleanup (to eliminate unnecessary conversions).

Note this is a breaking change and a fix will be submitted with the roll.

Bug: b/239435405
Change-Id: Ie5de1bcd0c2ec40450b154f5b36e26b3bde4579d