[Open Screen] Replace ByteView with a generic Span type.
This replaces ByteView with a generic Span type that is generally
forward compatible with C++20 std::span.  Span permits a view on both
const and non-const uint8_t data.

ByteView becomes an alias for Span<const uint8_t>.  ByteBuffer is an
alias for Span<uint8_t> and is used for APIs that mutate their input.

Updates the public APIs for libcast to use the aliases and removes
remaining dependencies in public headers on absl::Span.

Removing absl::Span from the remainder of the codebase can be done as a
mechanical cleanup (to eliminate unnecessary conversions).

Bug: b/239435405
Change-Id: Ie5de1bcd0c2ec40450b154f5b36e26b3bde4579d