Do not pass EVP_PKEY by value
As is standard for C APIs, it is unsafe and undefined to simply copy the
struct with memcpy. This violates all internal invariants in the library
and was only working by accident.

In particular, it would be a reasonable optimization for X509_set_pubkey
to bump the reference count to the EVP_PKEY and retain a pointer to it.
Upstream OpenSSL does just that. Had that been the case, openscreen's
invalid use of the API would have caused a stack use-after-free and
likely exploitable security vulnerability.

A recent change in BoringSSL, to make the private EVP_PKEY type opaque,
finally flagged this bug, but this was always unsafe and unsupported.

Alas, some const_casts are needed because the APIs are non-const to
allow them to bump the refcount (the thing that makes openscreen's usage
unsafe), and C's type system isn't capable enough to express all the
different flavors of const we have here. (Probably the right type would have
been shared_ptr<const EVP_PKEY>, but we don't have that here.)

While I'm here, fix some comments. The const_cast in i2d_X509 is
actually subtle, until we fix an issue, so I've linked to the BoringSSL
bug. The comment that BoringSSL doesn't free the buffer for you makes no
sense because the function inherently cannot free it, so rephrase it to
reflect the documented API contract.

Change-Id: I285d31e801b790fd62cf2c46f50e79dfc129762d