[Open Screen] Replace ConstDataSpan with ByteView.
ConstDataSpan was a struct used to represent a pointer-and-length in
the libcast authentication code.  Replace it with ByteView which serves
the same purpose.

Also adds some conversion functions and a const std::vector to Span
to simplify ByteView conversions to/from various other objects.  This
allows removal of various inline casts and conversion macros.

Eventually I want to remove all uses of std::string and
pointer-and-length for binary data in this code, but this is a step
in that direction.

Change-Id: Idce296374e7382baf8d45927da60422676df635e