Fix ASAN crash, enable OSP unittests on Mac
This patch fixes an ASAN crash caused by accessing a service info
after it has been logically removed but not deleted yet. This patch
also contains changes to enable these tests on Mac OSX (the platform
I resolved this bug on).

For an example failure, see:
https://chromium-review.googlesource.com/c/chromium/src/+/4603813

To reproduce locally, just build and run the unittests executable
with the GN arg "is_asan=true". It is a bummer that this isn't currently
caught with the Open Screen bots, only the Chromium bots.

Change-Id: I6a7cd5c87be17dc8b211ed3cb43219de8b1562c3