[Open Screen] Bulk-update license headers.
The license header presubmit check enforces the new license header
format which removes the string "All Rights Reserved."  Chromium was
updated but not our repo, so doing this now so that unrelated changes
aren't blocked on license headers.

To pass cpplint, this also fixes a number of IWYU issues and also
adds `noexcept` to move operators.

Bug: b/288403513
Change-Id: I90b2e58dcc2bdd1762129c7b5894e3a2730da51b