[Open Screen] Remove SerialDeletePtr from platform API.
This convenience class attempts to customize std::unique_ptr, which is
non-idiomatic and not something we want library clients to know about.

This CL removes SerialDeletePtr from the public API and converts
public usages to use std::unique_ptr<> with a custom deleter, similar
to Chromium's base:.

scoped_wake_lock also uses SerialDeletePtr and will need to be
migrated in a followup change.

Bug:
Change-Id: I9715790c471ad7334f6e9a186075dd40fc27ff15