[Open Screen] Remove SerialDeletePtr from platform API.
This convenience class attempts to customize std::unique_ptr, which is
non-idiomatic and not something we want library clients to know about.

This CL removes SerialDeletePtr from the public API and converts
public usages to use std::unique_ptr<> with a custom deleter, similar
to Chromium's base:.

Eventually we may remove SerialDeletePtr entirely, but there are many
internal usages.

Bug: b/288327294
Change-Id: I9715790c471ad7334f6e9a186075dd40fc27ff15