[ssci] Added Shipped field to READMEs
This CL adds the Shipped field (and may update the
License File field) in Chromium READMEs. Changes were
automatically created, so if you disagree with any of
them (e.g. a package is used only for testing purposes
and is not shipped), comment the suggested change and
why.

See the LSC doc at go/lsc-chrome-metadata.

Additionally, the script tools/licenses.py was manually
modified to better match the current state of Chromium's
//tools/licenses/licenses.py script. Namely,
    * prefer the Shipped field
    * log warnings for the deprecated
      NOT_SHIPPED value in the License File field
    * handle multiple license files being specified
    * paths to template files for GenerateCredits were
      updated so that credits will run if the openscreen repo
      is within a chromium/src checkout.

Bug: b:285450740
Change-Id: I0e838803d216a6d334ff69b0d9b049823f648137