[Cast Streaming] Add new StatisticsCollector class
This patch is focused on adding statistics collection to the Cast
Streaming library. These statistics are generated as part of creating
frames and packets, and then collated in a new StatisticsCollector
class.

Next up: take the collected statistics events and analyze them in order
to plumb generated statistics to the existing StatisticsClient API.

https://chromium-review.googlesource.com/c/openscreen/+/4583037

Bug: b/285417419
Change-Id: I66d1639ade649a538fae86b7c1acfcd100324668