[OpenScreen] Remove third_party/mozilla
Gooogleurl has already been ported into OpenScreen in [1], which
provides all the features provided by third_party/mozilla. So we can
use googleurl directly and there is no need to keep third_party/mozilla.

A special case is: url::CompareSchemeComponent can not be used in
util/url.cc, becuase it is not marked as COMPONENT_EXPORT in chromium.
An alternative method is added and used in util/url.cc.

[1] https://chromium-review.googlesource.com/c/openscreen/+/4730307

Bug: none
Change-Id: Ib8f9b3747eb92e622ea58fedbd1537b672081056