[Open Screen] Make receiver_message.h a public header.
This is part of the API for SenderSessionMessenger which is used in
Chromium.  Keeping it private prevents downstream from including it
properly in the client's implementation of the RpcDispatcher.
See crrev.com/c/5081140 for details.

Arguably, the entire RpcDispatcher could be moved from Chromium to
libcast, as it is really just a higher level API around the
SenderSessionMessenger.  Or, clients could use the SSM directly.

For now I want to land this to unblock landing crrev.com/c/5081140.

Change-Id: I59fd1cc4a46558abc04e2d353691b0924a3ce814