[Openscreen] Enable standalone cast_receiver target in Chromium
This target will be used for cast mirroring testing on ChromeOS.

There is a limitation however: building with optional dependencies, like
libvpx for decoding has not been enabled for Chromium builds.

Some of the dependencies could be provided by the Chrome sysroot
and statically linked into the resulting cast_receiver binary.

Bug: b/320187114
Change-Id: I412425d894672688e82855a4570a7eb496355a3d