Enable standalone cast_receiver target in Chrome build
This target will be used for cast mirroring testing on ChromeOS.

There is a limitation that will be addressed in a later CL: building
with optional dependencies, like libvpx for decoding, has not been
tested. Some of the dependencies could be provided by the Chrome sysroot
and statically linked into the resulting cast_receiver binary.

Bug: b:20187114
Test: Successfully built cast_receiver as a Chrome target.
Change-Id: I412425d894672688e82855a4570a7eb496355a3d