[OpenScreen] Replace QUIC with QUICHE to re-enable osp_demo
Currently, the QUIC implementaion in OSP targets is base on a fork
from Chromium. It is hard to maintain  and is deleted. As a result,
the osp_demo is not workable any more.

This CL replaces old QUIC implementation with QUICHE, which is a
portable implementation of QUIC and is much easier to maintain.
With this change, osp_demo becomes workable again.

Bug: b/169447969
Change-Id: I6871bc3ca238a2c5bb861eda6cf289033ef096d6