[Open Screen] Fix crash in Sender::EnqueueFrame()
Senders insist on providing frames with invalid reference times that are
non-montonically increasing.  This is a client bug, but the main
consequence is inaccurate sender reports that are used by the receiver
for A/V sync, so we carry on instead of crashing.

NOTE: Attempts to write a fine gained unit test were unsuccessful, as
various layers of the stack make timestamp conversions that actually
change underlying sender report values, making it hard to write
easy-to-understand expectations on the sender reports.  This would be
worth further investigation.

Bug: 330810255
Change-Id: Ib2db10ed5a470f4e31a4a206e09cc9d5d772c614