[OpenScreen] Save a copy of Config in MdnsServiceImpl
Currently Config is created in CreateDnsSdServiceInternal function as
a local variable. However, it is referenced in some classes, such as
MdnsResponder. This may cause use-after-free issue. Besides, it is
copied in some classes, such as MdnsQuerier. This is unnecessary.

This CL fixes the issues by saving a copy of Config in MdnsServiceImpl
and then using const reference everywhere.

Bug: none
Change-Id: Icdc8a2352f2b7e659f4f77a34cf76e7bdcb668a9