[OpenScreen] Include agent fingerprint in DNS TXT records
According to the spec[1], the agent fingerprint of the advertising
agent should be included in DNS TXT records.

Once this CL is merged, we can generate agent certificate on the
server side and send the agent fingerprint to client side. Then
client can use the agent fingerprint to verify agent certificate.

[1] https://w3c.github.io/openscreenprotocol/#discovery

Bug: none
Change-Id: Ia0248e29f765973f4d7fc0236c4511bba3d4137f