[Open Screen] Fix TODOs.
This CL removes some obsolete TODOs and fixes a remaining TODO.

The fix is to move the open screen protocol message definitions from
the cast:: namespace to the openscreen::proto:: namespace, to be
consistent with the rest of the code in the library, which is all
in the openscreen:: namespace.

This results in a large number of mechanical changes elsewhere.  A
corresponding change will need to be landed in Chromium with the
DEPS roll.

Bug: 155336728
Change-Id: I8fbc78d3c50265953b9dc0b15241e438a64d5195