[Open Screen] Simplify is_debug builds.
is_debug = true is the default for builds.  Update the LUCI
configuration to set it only when false and to properly
configure msan builds.

This also fixes some msan failures that were hidden by the fact
that we were not running msan in the CI/CQ.  That is fixed
by crrev.com/c/5647554

Bug: 347732145
Change-Id: Iec167e3d3e30628492e316f70fa3c870f4cfa401