[OpenScreen] Make MessageDemuxer be a member of QuicServiceBase
QuicServiceBase uses MessageDemuxer for processing received data, so the
life time of MessageDemuxer should not be shorter than QuicServiceBase.
However, this can not be guaranteed now, because MessageDemuxer is saved
as a reference in QuicServiceBase. This CL makes MessageDemuxer be a
member of QuicServiceBase to guarantee that. This CL also changes the
interfaces of FakeQuicBridge as a result of this change.

Bug: none
Change-Id: Icb2eac41c009ff487a83152646880772fe9fdc7a