[Openscreen] Let standalone_impl be built from Chromium
This should be a no-op change. The source set cannot be included
directly from Chromium. Furthermore, it is not visible to any
source sets other than some standalone targets. Currently, none of these
are visible from Chromium.

This also adds `testonly = true` to the platform:standalone_impl source
set when it is built from Chromium. This will ensure it is not
accidentally included in Chromium.

Bug: b/320187114
Change-Id: Ie31150a4492597539fe56b35d24cd9e031b7f61e