[OpenScreen] reduce standalone_receiver dependence platform/impl
This should have no effect.

cast/standalone_receiver/streaming_playback_controller.h depends on the
TaskRunner base class in platform/api, not on the implementation in
platform/impl. Switching the include to use the API definition is
cleaner.

Actual dependence on platform/impl is isolated to main.cc. Reducing the
build dependence helps prepare for constraining the visibility of the
platform/impl source sets to binary test targets.

Bug: b/320187114
Change-Id: I086c4f777c41b01dcd30611f4beb059d19e15fb6