[OpenScreen] Improve move semantic of MessageWatch
Move semantic of MessageWatch is confusing now: if we move an invalid
MessageWatch to a valid MessageWatch, the invlid one will become valid.
However, move semantic generally means that the object is left in an
invalid state after being moved. This CL improves semantic of
MessageWatch and adds Reset() interface.

Bug: none
Change-Id: Ifb282dac2c252bd766106ef2d1457ac75377d034