[OpenScreen] Improve move semantic for ConnectRequest and ReceiverWatch
Their move semantic is confusing now: if we move an invalid value to
a valid value, the invlid one will become valid. However, move
semantic generally means that the object is left in an invalid
state after being moved. This CL improves move semantic for them
to make it clear.

Bug: 281741443
Change-Id: I6f5a04bbaa2155a2240d561d60008aa6bc052f3a