[Openscreen] Add standalone cast_receiver target to Chrome builds
This adds the standalone cast_receiver target to non-Windows builds of
Chrome. A similar change in https://crrev.com/c/5369648 was reverted due
to issues when building the target for Windows or Android.

This limits enabling the cast_receiver target to only non-Windows
builds. The Android issue should be fixed by allowing Android builds to
use Linux network adapter enumeration.

Bug: b/320187114
Change-Id: I2d6476df8b486489ce17be3324b6013274174c1e